requirements.txt
websockets
flask
opencv-python
RPi.GPIO
